Implementation of Verification Process prior to starting a video call
-Implemented verification key into 'Join' link in email_data.php.
-Created function 'verifyKeys' in includes/songcircle.php which compares
a user provided verification key with one in the database.
-Included 'verification_key' in SQL query in
includes/songcircle_cronjobs.php to be able to pass this parameter to
constructHTMLEmail function.
-Updated confirmUserRegistration function in
includes/songcircle_user_action case: confirm_registration to insert
verification key into songcircle_register table when a user confirms for
a songcircle
-Wrote validation for query string to start_call.php to check for
presence of parameters and to compare and qualify verification keys
-Added 'verification_key' column to table songcircle_register